Move files out of models/
