+++ src/contributions/editors/stbl-binary/document.ts
@@ -4,7 +4,7 @@ import { StringTableResource } from '@s4tk/models';
 import { StringTableLocale } from '@s4tk/models/enums';
 import { formatAsHexString } from '@s4tk/hashing/formatting';
 import { S4TKSettings } from '#helpers/settings';
-import StringTableJson from '#models/stbl-json';
+import StringTableJson from '#stbls/stbl-json';
 import { parseKeyFromTgi } from '#building/resources';
 import ViewOnlyDocument from '../view-only/document';
